requirements.txt
Flask
openai
python-dotenv
paramiko
psutil